requirements.txt
streamlit
pandas
langchain
openai
llama-index
PyMuPDF
python-docx
tiktoken
faiss-cpu
